+++ src/Components/Navbar/Navbar.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "./Navbar.css";
 import { WiMoonWaningCrescent3 } from "react-icons/wi";
-const Navbar = () => {
+export const Navbar = () => {
   return (
     <div className="navbar">
       <h1 className="navbar-heading">Notes App</h1>
@@ -11,4 +11,4 @@ const Navbar = () => {
   );
 };
 
-export default Navbar;
+
