+++ src/Components/Note/Note.css
@@ -1,3 +1,4 @@
+
 .notes {
   border: 1px solid black;
   width: 60%;
