+++ src/Components/index.js
@@ -1,0 +1,6 @@
+import { Navbar } from "./Navbar/Navbar"
+import { Sidebar } from "./Sidebar/Sidebar"
+import { SearchBar } from "./SearchBar/SearchBar"
+import { Note } from "./Note/Note"
+
+export {Navbar, Sidebar, SearchBar,Note}
