+++ src/pages/Profile/Profile.jsx
@@ -1,9 +1,8 @@
 import React from 'react'
 
-const Profile = () => {
+export const Profile = () => {
   return (
     <div>Profile</div>
   )
 }
 
-export default Profile
