+++ src/pages/Trash/Trash.jsx
@@ -1,9 +1,8 @@
 import React from 'react'
 
-const Trash = () => {
+export const Trash = () => {
   return (
     <div>Trash</div>
   )
 }
 
-export default Trash
