+++ src/pages/index.js
@@ -1,0 +1,7 @@
+import { Archive } from "./Archive/Archive";
+import { Trash } from "./Trash/Trash";
+import { Profile } from "./Profile/Profile";
+import { Homepage } from "./HomePage/Homepage";
+import { Labels } from "./Labels/Labels";
+
+export {Archive, Trash, Profile, Homepage, Labels}
